requirements.txt
dnspython==2.7.0
iniconfig==2.0.0
mysql-connector-python==9.1.0
packaging==24.2
pluggy==1.5.0
pymongo==4.10.1
pytest==8.3.3
